tests/Document-createProcessingInstruction-literal-1.xhtml

<?xml version="1.0" encoding="UTF-8"?>
<html xmlns="http://www.w3.org/1999/xhtml">
<head>
<title>&lt;?xml?> is not a ProcessingInstruction</title>
<link rel="author" title="Ms2ger" href="mailto:ms2ger@gmail.com"/>
<link rel="help" href="../web-dom-core#processinginstruction"/>
<script src="testharness.js"></script>
</head>
<body>
<div id="log"/>
<script>
test(function() {
  assert_equals(document.firstChild, document.documentElement)
})
</script>
</body>
</html>
